// src/backend/types/Muscle.ts
export enum Muscle {
  LowerBack = 'Lower Back',
  Soleus = 'Soleus',
  FrontDeltoids = 'Front Deltoids',
  Lats = 'Lats',
  Forearms = 'Forearms',
  Pecs = 'Pecs',
  Hamstrings = 'Hamstrings',
  WristFlexors = 'Wrist Flexors',
  Biceps = 'Biceps',
  Triceps = 'Triceps',
  RearDeltoids = 'Rear Deltoids',
  RotatorCuff = 'Rotator Cuff',
  Ankle = 'Ankle',
  Abdominals = 'Abdominals',
  Glutes = 'Glutes',
  Quadriceps = 'Quadriceps',
  Obliques = 'Obliques',
  Abductors = 'Abductors',
  Gastrocnemius = 'Gastrocnemius',
  LateralDeltoids = 'Lateral Deltoids',
  HipFlexors = 'Hip Flexors',
  Trapezius = 'Trapezius',
  Neck = 'Neck',
  Adductors = 'Adductors',
}